refactor: IssueDetailHeader 컴포넌트 분리
- ghojeong/issue-tracker/#183
