+++ frontend/issue-cracker/src/components/layout/IssueDetail/IssueDetailHeaderTitleSection.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import IssueDetailTitleEdit from './IssueDetailTitleEdit';
+import IssueDetailTitle from './IssueDetailTitle';
+
+const IssueDetailHeaderTitleSection = ({
+  state,
+  content,
+}: {
+  state: boolean;
+  content: string;
+}): JSX.Element => {
+  return (
+    <>
+      {state ? <IssueDetailTitleEdit /> : <IssueDetailTitle {...{ content }} />}
+    </>
+  );
+};
+
+export default IssueDetailHeaderTitleSection;
